--- a/cactusref/README.md
+++ b/cactusref/README.md
@@ -53,7 +53,7 @@ to all nodes in the cycle come from other nodes in the cycle.
 Cycle detection is a zero-cost abstraction. If you never
 `use cactusref::Adoptable;`, `Drop` uses the same implementation as
 `std::rc::Rc` (and leaks in the same way as `std::rc::Rc` if you form a cycle of
-strong references). The only costs you pay are the memory costs of two empty
+strong references). The only costs you pay are the memory costs of one empty
 [`RefCell`](https://doc.rust-lang.org/nightly/core/cell/struct.RefCell.html)`<`[`HashMap`](https://docs.rs/hashbrown/0.5.0/hashbrown/struct.HashMap.html)`<NonNull<T>, usize>>`
 for tracking adoptions and an if statement to check if these structures are
 empty on `drop`.