--- a/cactusref/src/adoptable.rs
+++ b/cactusref/src/adoptable.rs
@@ -126,7 +126,7 @@ unsafe impl<T: ?Sized> Adoptable for Rc<T> {
         // Remove a forward reference to `other` in `this`. This bookkeeping
         // logs a strong reference and is used for discovering cycles.
         let mut links = this.inner().links.borrow_mut();
-        links.remove(Link::forward(other.ptr));
+        links.remove(Link::forward(other.ptr), 1);
         // `this` and `other` may be the same `Rc`. Drop the borrow on `links`
         // before accessing `other` to avoid a already borrowed error from the
         // `RefCell`.
@@ -134,6 +134,6 @@ unsafe impl<T: ?Sized> Adoptable for Rc<T> {
         // Remove a backward reference to `this` in `other`. This bookkeeping is
         // used for discovering cycles.
         let mut links = other.inner().links.borrow_mut();
-        links.remove(Link::backward(this.ptr));
+        links.remove(Link::backward(this.ptr), 1);
     }
 }