--- a/cactusref/src/link.rs
+++ b/cactusref/src/link.rs
@@ -2,7 +2,13 @@ use core::ptr::{self, NonNull};
 use hashbrown::{hash_map, HashMap};
 use std::hash::{Hash, Hasher};
 
-use crate::ptr::RcBox;
+use crate::ptr::{RcBox, RcBoxPtr};
+
+#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq)]
+pub enum Kind {
+    Forward,
+    Backward,
+}
 
 pub struct Links<T: ?Sized> {
     pub registry: HashMap<Link<T>, usize>,
@@ -19,7 +25,7 @@ impl<T: ?Sized> Links<T> {
 
     pub fn remove(&mut self, other: Link<T>) {
         match self.registry.get(&other).copied().unwrap_or_default() {
-            0 | 1 => self.registry.remove(&other),
+            count if count <= 1 => self.registry.remove(&other),
             count => self.registry.insert(other, count - 1),
         };
     }
@@ -30,10 +36,6 @@ impl<T: ?Sized> Links<T> {
 
     pub fn is_empty(&self) -> bool {
         self.registry.is_empty()
-    }
-
-    pub fn len(&self) -> usize {
-        self.registry.len()
     }
 
     pub fn iter(&self) -> hash_map::Iter<Link<T>, usize> {
@@ -57,19 +59,59 @@ impl<T: ?Sized> Default for Links<T> {
     }
 }
 
-pub struct Link<T: ?Sized>(pub NonNull<RcBox<T>>);
+// Using a a tuple struct is about 10% faster than using named fields.
+pub struct Link<T: ?Sized>(NonNull<RcBox<T>>, Kind);
+
+impl<T: ?Sized> Link<T> {
+    #[inline]
+    pub fn forward(ptr: NonNull<RcBox<T>>) -> Self {
+        Self(ptr, Kind::Forward)
+    }
+
+    #[inline]
+    pub fn backward(ptr: NonNull<RcBox<T>>) -> Self {
+        Self(ptr, Kind::Backward)
+    }
+
+    #[inline]
+    pub fn link_kind(&self) -> Kind {
+        self.1
+    }
+
+    #[inline]
+    pub fn as_forward(&self) -> Self {
+        Self::forward(self.0)
+    }
+
+    #[inline]
+    pub fn as_ptr(&self) -> *const RcBox<T> {
+        self.0.as_ptr()
+    }
+
+    #[inline]
+    pub fn into_raw_non_null(self) -> NonNull<RcBox<T>> {
+        self.0
+    }
+}
+
+impl<T: ?Sized> RcBoxPtr<T> for Link<T> {
+    #[inline]
+    fn inner(&self) -> &RcBox<T> {
+        unsafe { self.0.as_ref() }
+    }
+}
 
 impl<T: ?Sized> Copy for Link<T> {}
 
 impl<T: ?Sized> Clone for Link<T> {
     fn clone(&self) -> Self {
-        Self(self.0)
+        Self(self.0, self.1)
     }
 }
 
 impl<T: ?Sized> PartialEq for Link<T> {
     fn eq(&self, other: &Self) -> bool {
-        ptr::eq(self.0.as_ptr(), other.0.as_ptr())
+        ptr::eq(self.0.as_ptr(), other.0.as_ptr()) && self.1 == other.1
     }
 }
 
@@ -78,5 +120,6 @@ impl<T: ?Sized> Eq for Link<T> {}
 impl<T: ?Sized> Hash for Link<T> {
     fn hash<H: Hasher>(&self, state: &mut H) {
         self.0.hash(state);
+        self.1.hash(state);
     }
 }