--- a/cactusref/tests/no_leak_doubly_linked_list.rs
+++ b/cactusref/tests/no_leak_doubly_linked_list.rs
@@ -84,18 +84,20 @@ fn cactusref_doubly_linked_list_no_leak() {
     env_logger::Builder::from_env("CACTUS_LOG").init();
 
     // 500MB of `String`s will be allocated by the leak detector
-    leak::Detector::new("CactusRef adopt self", ITERATIONS, LEAK_TOLERANCE).check_leaks(|_| {
-        let list = iter::repeat(())
-            .map(|_| "a".repeat(1024 * 1024))
-            .take(10)
-            .collect::<Vec<_>>();
-        let mut list = List::from(list);
-        let head = list.pop().unwrap();
-        assert_eq!(Rc::strong_count(&head), 1);
-        assert_eq!(list.head.as_ref().map(Rc::strong_count), Some(3));
-        let weak = Rc::downgrade(&head);
-        drop(head);
-        assert!(weak.upgrade().is_none());
-        drop(list);
-    });
+    leak::Detector::new("CactusRef doubly linked list", ITERATIONS, LEAK_TOLERANCE).check_leaks(
+        |_| {
+            let list = iter::repeat(())
+                .map(|_| "a".repeat(1024 * 1024))
+                .take(10)
+                .collect::<Vec<_>>();
+            let mut list = List::from(list);
+            let head = list.pop().unwrap();
+            assert_eq!(Rc::strong_count(&head), 1);
+            assert_eq!(list.head.as_ref().map(Rc::strong_count), Some(3));
+            let weak = Rc::downgrade(&head);
+            drop(head);
+            assert!(weak.upgrade().is_none());
+            drop(list);
+        },
+    );
 }