Performance is all about trade-offs. This patch is not universally better.

Pros:
- Cycle detection algorithm is significantly simpler to understand
- Link traversal in drop cases is simpler to understand
- Adoption bookkeeping is more explicit and simpler to understand
- `RcBox` size has decreased from 112 bytes to 64 bytes

Cons:
- Cycle detection is slower.
- The performance degradation is worse for sparsely connected graphs. About 15% slower for a singly linked ring.

5% slower in the worst case complexity + simpler implementation + half memory usage for the no-adoption case is a trade-off I am willing to make.

Bench run:

```console
$ cargo bench -p cactusref
drop a circular graph/10
                        time:   [21.325 us 21.643 us 21.992 us]
                        change: [+14.904% +17.185% +19.838%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) low mild
  4 (4.00%) high mild
  1 (1.00%) high severe
drop a circular graph/20
                        time:   [59.318 us 59.983 us 60.670 us]
                        change: [+13.297% +14.917% +16.485%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe
drop a circular graph/30
                        time:   [126.09 us 127.25 us 128.41 us]
                        change: [+13.567% +15.032% +16.431%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe
drop a circular graph/40
                        time:   [180.68 us 181.89 us 183.12 us]
                        change: [+12.199% +14.038% +16.072%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 8 outliers among 100 measurements (8.00%)
  2 (2.00%) high mild
  6 (6.00%) high severe
drop a circular graph/50
                        time:   [247.86 us 249.77 us 251.66 us]
                        change: [+11.269% +12.802% +14.489%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  5 (5.00%) high mild
drop a circular graph/100
                        time:   [802.12 us 806.81 us 811.70 us]
                        change: [+10.868% +12.924% +14.989%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 8 outliers among 100 measurements (8.00%)
  8 (8.00%) high severe

drop a fully connected graph/10
                        time:   [49.171 us 49.746 us 50.443 us]
                        change: [+6.6333% +8.2439% +9.8562%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
drop a fully connected graph/20
                        time:   [238.42 us 240.57 us 242.95 us]
                        change: [+5.7385% +7.0547% +8.2935%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe
drop a fully connected graph/30
                        time:   [687.28 us 692.42 us 698.23 us]
                        change: [+1.6952% +2.8716% +4.1074%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 11 outliers among 100 measurements (11.00%)
  2 (2.00%) low mild
  4 (4.00%) high mild
  5 (5.00%) high severe
drop a fully connected graph/40
                        time:   [1.4594 ms 1.4698 ms 1.4809 ms]
                        change: [-0.0632% +1.3767% +3.1430%] (p = 0.10 > 0.05)
                        No change in performance detected.
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) high mild
  3 (3.00%) high severe
drop a fully connected graph/50
                        time:   [2.8164 ms 2.8408 ms 2.8654 ms]
                        change: [+4.0132% +4.9724% +5.9184%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
drop a fully connected graph/100
                        time:   [21.931 ms 22.026 ms 22.134 ms]
                        change: [-1.0727% +0.2455% +1.4751%] (p = 0.71 > 0.05)
                        No change in performance detected.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe
```